Moved ProjectVisibility to graph-commons
