+++ webhook-service/app/ch/datascience/webhookservice/project/ProjectInfoFinder.scala
@@ -21,9 +21,10 @@ package ch.datascience.webhookservice.project
 import cats.effect.IO
 import ch.datascience.http.client.{AccessToken, IORestClient}
 import ch.datascience.graph.events._
+import ch.datascience.graph.model.project.ProjectVisibility
 import ch.datascience.webhookservice.IOContextShift
 import ch.datascience.webhookservice.config.IOGitLabConfigProvider
-import ch.datascience.webhookservice.model.{ProjectInfo, ProjectOwner, ProjectVisibility}
+import ch.datascience.webhookservice.model.{ProjectInfo, ProjectOwner}
 import javax.inject.{Inject, Singleton}
 
 import scala.concurrent.ExecutionContext
