+++ webhook-service/test/ch/datascience/webhookservice/generators/ServiceTypesGenerators.scala
@@ -20,6 +20,7 @@ package ch.datascience.webhookservice.generators
 
 import ch.datascience.generators.Generators._
 import ch.datascience.graph.events.EventsGenerators._
+import ch.datascience.graph.events.GraphCommonsGenerators._
 import ch.datascience.webhookservice.crypto.HookTokenCrypto.SerializedHookToken
 import ch.datascience.webhookservice.eventprocessing.PushEvent
 import ch.datascience.webhookservice.model._
@@ -50,8 +51,6 @@ object ServiceTypesGenerators {
   implicit val projectOwner: Gen[ProjectOwner] = for {
     id <- userIds
   } yield ProjectOwner(id)
-
-  implicit val projectVisibilities: Gen[ProjectVisibility] = Gen.oneOf(ProjectVisibility.all.toList)
 
   implicit val projectInfos: Gen[ProjectInfo] = for {
     id         <- projectIds
