+++ webhook-service/test/ch/datascience/webhookservice/hookvalidation/HookValidatorSpec.scala
@@ -26,11 +26,11 @@ import ch.datascience.generators.Generators.Implicits._
 import ch.datascience.generators.Generators.exceptions
 import ch.datascience.graph.events.GraphCommonsGenerators._
 import ch.datascience.graph.events.ProjectId
+import ch.datascience.graph.model.project.ProjectVisibility
 import ch.datascience.interpreters.TestLogger
 import ch.datascience.interpreters.TestLogger.Level.{Error, Info}
 import ch.datascience.webhookservice.generators.ServiceTypesGenerators._
 import ch.datascience.webhookservice.hookvalidation.HookValidator.HookValidationResult.{HookExists, HookMissing}
-import ch.datascience.webhookservice.model.ProjectVisibility
 import ch.datascience.webhookservice.project._
 import io.chrisdavenport.log4cats.Logger
 import org.scalacheck.Arbitrary
